Significant changes / enhancements / rationalisations / optimisations for the management of GpuProgramParameters - The underlying storage of parameters is no longer locked to separate entries of float4/int4. Instead, there is a raw float / int buffer holding all the parameter values and the allocation of sections of it is done based on uniform definitions (for high-level parameters) or constant setting (for low-level parameters) - There are no longer any unused gaps in the parameter storage (mapping to predefined constants in the shader), so no need to trace 'isSet' values. - You can grab a 'physical' buffer index from either a parameter name (high level) or logical index (low-level) and then use that index to set entries in the parameter buffer directly at runtime without the lookup overhead - D3D9 can now set a parameter arrays in one call instead of looping - All low-level programs still use the underlying '4 float rule' ie all constant entries are at least 4 elements in size. - HLSL and Cg both map down to assembler which still obeys the 4-float rule - GLSL does not obey the 4-float rule and parameter data can be packed arbitrarily in our buffers. We can now set GLSL parameters without having to copy the values out of 4-float constants into a packed buffer since the structure is already correct. note, some care is needed now - don't declare a "uniform vec3 colour" and then auto-bind a colour to it, since that's 4 floats and the remaining float will overflow. We do (overall) bounds checking in debug mode which catches this but only at the end of a buffer. - Named parameter lookups are now held in the HighLevelGpuProgram instance and shared among all parameter objects to save space - GpuProgramParameters::ConstantDefinition and associated parameters removed. Replaced with GpuConstantDefinition (non-member struct) which serves a dual purpose; it maps out the buffer structure, it creates name/buffer mappings, and it records the type, element size and array size. There is now no need for serialiser or user code to call addConstantDefinition to define constants, just loading a program or creating a params object will do it. This information is held on the GpuProgram/HighLevelGpuProgram and shared with parameter objects to reduce the size - Re-wrote large portions of the GLSL parameter code, it should now be much faster since it no longer has to perform parameter name matching. Mappings are constructed during uniform construction (which uses source code now) and re-used at parameter setting time. - Some fixes to the GLSL uniform derivation code; now correctly ignores commented out uniform definitions and use of the word 'uniform' as a sub-part of something different. Also parses attached sources as well in case uniforms are declared there. - Rationalised the use of array indices. You now no longer have to use "[0]" at the end of the name of an array uniform, although you can do if you want. All high-level programs obey the same rules, and for an array parameter 'param' you can use 'param', 'param[0]',and if the array is up to 16 entries large, all the individual indexes up to its limit. Larger arrays are not individually indexed beyond [0] since it would clutter the name map, and we expect you to set the parameters in bulk rather than indexing each individually. -  GpuProgramParameters::setNamedConstant(name, pointer, count) now takes an additional 'multiple' parameter so you can set values in non-float4 chunks if you want. This is required for GLSLprograms using arrays of structures that are smaller than 4 elements wide. param_named automatically deals with this. The default is 4 for compatibility with previous code.

+++ trunk/OgreMain/src/OgreFileSystem.cpp
@@ -78,8 +78,8 @@ namespace Ogre {
         struct _finddata_t tagData;
 
         // pattern can contain a directory name, separate it from mask
-        unsigned int pos1 = pattern.rfind ('/');
-        unsigned int pos2 = pattern.rfind ('\\');
+        size_t pos1 = pattern.rfind ('/');
+        size_t pos2 = pattern.rfind ('\\');
         if (pos1 == pattern.npos || ((pos2 != pattern.npos) && (pos1 < pos2)))
             pos1 = pos2;
         String directory;
